backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
psycopg2-binary==2.9.9
asyncpg==0.30.0
openpyxl==3.1.5
reportlab==4.4.1
alembic==1.13.1
greenlet==3.0.3

# Rate limiting
slowapi==0.1.9

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
aiosqlite==0.20.0

# Monitoring
sentry-sdk[fastapi]==1.40.6
